Add files via upload
Hi all, I took the time to do a little study of the aiming mechanics of halo 4 and infinite and attempted to recreate the halo 4 feel in infinite.  I made a pdf with figures that completely explains my methodology, my findings, and all that.  Please look through it if you care about this sort of thing, there's a lot that's different in halo infinite and I think its worth a read.